added context to colelction

+++ saleor/my_customs/merchant_dashboard/collection/views.py
@@ -8,6 +8,7 @@ from django.urls import reverse
 from django.utils.translation import pgettext_lazy
 from django.views.decorators.http import require_POST
 
+from saleor.my_customs.merchant.models import Merchant
 from ....core.utils import get_paginator_items
 from ....product.models import Collection
 from ..menu.utils import get_menus_that_needs_update, update_menus
@@ -19,6 +20,13 @@ from .forms import AssignHomepageCollectionForm, CollectionForm
 @staff_member_required
 @permission_required('product.manage_products')
 def collection_list(request):
+    user = request.user
+    if not user.is_authenticated:
+        return  # todo: return unauthorized page
+    merchant = Merchant.objects.get_merchant_of_user(user)
+    if not merchant:
+        return TemplateResponse(request, 'my_customs/merchant_dashboard/not_registered.html')
+
     site_settings = request.site.settings
     assign_homepage_col_form = AssignHomepageCollectionForm(
         request.POST or None, instance=site_settings)
@@ -27,8 +35,8 @@ def collection_list(request):
         msg = pgettext_lazy(
             'Dashboard message', 'Updated homepage collection')
         messages.success(request, msg)
-        return redirect('dashboard:collection-list')
-    collections = Collection.objects.prefetch_related('products').all()
+        return redirect('merchant_dashboard:collection-list')
+    collections = Collection.objects.prefetch_related('products').get_by_merchant(merchant)
     collection_filter = CollectionFilter(request.GET, queryset=collections)
     collections = get_paginator_items(
         collection_filter.qs, settings.DASHBOARD_PAGINATE_BY,
@@ -44,14 +52,21 @@ def collection_list(request):
 @staff_member_required
 @permission_required('product.manage_products')
 def collection_create(request):
+    user = request.user
+    if not user.is_authenticated:
+        return  # todo: return unauthorized page
+    merchant = Merchant.objects.get_merchant_of_user(user)
+    if not merchant:
+        return TemplateResponse(request, 'my_customs/merchant_dashboard/not_registered.html')
+
     collection = Collection()
     form = CollectionForm(
-        request.POST or None, request.FILES or None, instance=collection)
+        request.POST or None, request.FILES or None, instance=collection, merchant=merchant)
     if form.is_valid():
         form.save()
         msg = pgettext_lazy('Collection message', 'Added collection')
         messages.success(request, msg)
-        return redirect('dashboard:collection-list')
+        return redirect('merchant_dashboard:collection-list')
     ctx = {'collection': collection, 'form': form}
     return TemplateResponse(request, 'my_customs/merchant_dashboard/collection/detail.html', ctx)
 
@@ -59,14 +74,21 @@ def collection_create(request):
 @staff_member_required
 @permission_required('product.manage_products')
 def collection_update(request, pk=None):
-    collection = get_object_or_404(Collection, pk=pk)
+    user = request.user
+    if not user.is_authenticated:
+        return  # todo: return unauthorized page
+    merchant = Merchant.objects.get_merchant_of_user(user)
+    if not merchant:
+        return TemplateResponse(request, 'my_customs/merchant_dashboard/not_registered.html')
+
+    collection = get_object_or_404(Collection.objects.get_by_merchant(merchant), pk=pk)
     form = CollectionForm(
-        request.POST or None, request.FILES or None, instance=collection)
+        request.POST or None, request.FILES or None, instance=collection, merchant=merchant)
     if form.is_valid():
         collection = form.save()
         msg = pgettext_lazy('Collection message', 'Updated collection')
         messages.success(request, msg)
-        return redirect('dashboard:collection-update', pk=collection.pk)
+        return redirect('merchant_dashboard:collection-update', pk=collection.pk)
     is_unpublish_restricted = (
         collection == request.site.settings.homepage_collection and
         collection.is_published)
@@ -79,7 +101,14 @@ def collection_update(request, pk=None):
 @staff_member_required
 @permission_required('product.manage_products')
 def collection_delete(request, pk=None):
-    collection = get_object_or_404(Collection, pk=pk)
+    user = request.user
+    if not user.is_authenticated:
+        return  # todo: return unauthorized page
+    merchant = Merchant.objects.get_merchant_of_user(user)
+    if not merchant:
+        return TemplateResponse(request, 'my_customs/merchant_dashboard/not_registered.html')
+
+    collection = get_object_or_404(Collection.objects.get_by_merchant(merchant), pk=pk)
     if request.method == 'POST':
         menus = get_menus_that_needs_update(collection=collection)
         collection.delete()
@@ -88,9 +117,9 @@ def collection_delete(request, pk=None):
         msg = pgettext_lazy('Collection message', 'Deleted collection')
         messages.success(request, msg)
         if request.is_ajax():
-            response = {'redirectUrl': reverse('dashboard:collection-list')}
+            response = {'redirectUrl': reverse('merchant_dashboard:collection-list')}
             return JsonResponse(response)
-        return redirect('dashboard:collection-list')
+        return redirect('merchant_dashboard:collection-list')
     ctx = {'collection': collection}
     return TemplateResponse(
         request, 'my_customs/merchant_dashboard/collection/confirm_delete.html', ctx)
@@ -100,7 +129,14 @@ def collection_delete(request, pk=None):
 @staff_member_required
 @permission_required('product.manage_products')
 def collection_toggle_is_published(request, pk):
-    collection = get_object_or_404(Collection, pk=pk)
+    user = request.user
+    if not user.is_authenticated:
+        return  # todo: return unauthorized page
+    merchant = Merchant.objects.get_merchant_of_user(user)
+    if not merchant:
+        return TemplateResponse(request, 'my_customs/merchant_dashboard/not_registered.html')
+
+    collection = get_object_or_404(Collection.objects.get_by_merchant(merchant), pk=pk)
     collection.is_published = not collection.is_published
     collection.save(update_fields=['is_published'])
     return JsonResponse(
